pyproject.toml
[project]
name = "smart-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.13.1",
    "markdown2[all]>=2.5.2",
    "markdown>=3.7",
    "openai>=1.58.1",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pydub>=0.25.1",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.41.1",
    "youtube-transcript-api>=0.6.3",
    "yt-dlp>=2024.12.13",
]
